screenshot: move to game/
