+++ src/meson.build
@@ -618,6 +618,7 @@ common_sources = [
   'trx/game/savegame/file.c',
   'trx/game/savegame/file_read.c',
   'trx/game/savegame/file_write.c',
+  'trx/game/screenshot.c',
   'trx/game/shell/args.c',
   'trx/game/shell/common.c',
   'trx/game/shell/config.c',
@@ -719,7 +720,6 @@ common_sources = [
   'trx/json/write.c',
   'trx/log.c',
   'trx/memory.c',
-  'trx/screenshot.c',
   'trx/strings/case_funcs.c',
   'trx/strings/common.c',
   'trx/strings/fuzzy_match.c',
