+++ src/trx/game/console/cmd/screenshot.c
@@ -3,7 +3,7 @@
 #include <trx/game/console/registry.h>
 #include <trx/game/game_string.h>
 #include <trx/game/output.h>
-#include <trx/screenshot.h>
+#include <trx/game/screenshot.h>
 #include <trx/strings.h>
 
 static COMMAND_RESULT M_Entrypoint(const COMMAND_CONTEXT *ctx)
