+++ src/trx/game/screenshot.h
